seprate source file into two
